tests/wpt/tests/scroll-animations/animation-trigger/animation-trigger-animation-target-removed-listener.tentative.html

<!DOCTYPE html>
<html>
<head>
  <link rel="help" src="https://drafts.csswg.org/css-animations-2/#animation-trigger">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/web-animations/testcommon.js"></script>
    <script src="/dom/events/scrolling/scroll_support.js"></script>
    <script src="support/support.js"></script>
</head>

<body>
  <style>
    .scroller {
      overflow-y: scroll;
      border: solid 1px;
      justify-self: center;
      height: 100%;
      width: 100%;
      display: inline-block;
    }

    .subject {
      height: 100px;
      width: 100%;
      background-color: black;
    }

    .space {
      height: 300px;
      width: 90%;
      justify-self: center;
    }

    #triggergrid {
      timeline-scope: --viewtimeline;
      height: 85vh;
      width: 90vw;
      display: grid;
      grid-template-columns: 1fr 1fr;
      .target {
        align-self: center;
        justify-self: center;
        height: 100px;
        width: 100px;
        background-color: blue;
      }

      .space {
        height: 130vh;
      }
    }
  </style>
  <div id="triggergrid">
    <div id="target" class="target"></div>
    <div id="test"></div>
    <div id="scroller" class="scroller">
      <div class="space"></div>
      <div class="space"></div>
      <div id="subject" class="subject"></div>
      <div class="space"></div>
      <div class="space"></div>
    </div>
  </div>
  <script>
    const scroller = document.getElementById("scroller");
    const subject = document.getElementById("subject");
    const ANIMATION_DURATION_MS = 1;

    async function attachFinishListener() {
      return new Promise(resolve => {
        const test = document.getElementById("test");
        const target = document.getElementById("target");
        const animation = target.getAnimations()[0];

        animation.addEventListener("finish", () => {
          test.innerText = "finish listener ran";
          resolve();
        }, { once: true });
      });
    }

    promise_test(async (test) => {
      waitForAnimationFrames(2);
      setupAnimationAndTrigger(document.getElementById("target"), subject,
        ANIMATION_DURATION_MS);
      const finish_promise = attachFinishListener();
      const animation = document.getElementById("target").getAnimations()[0];

      // Remove the animation's target. The animation's target may have been
      // removed, but since the animation can still be triggered and it has an
      // event listener, that event listener should fire when the trigger plays
      // the animation.
      document.getElementById("target").remove();

      // Scroll into the trigger range, triggering the animation.
      const contain_0_offset = computeContainOffset(scroller, subject, 0);
      await waitForScrollReset(test, scroller, 0, contain_0_offset);
      await finish_promise;

      const test_div = document.getElementById("test");
      assert_equals(test_div.innerText, "finish listener ran");

      assert_times_equal(animation.currentTime, ANIMATION_DURATION_MS,
        "animation currentTime updated");
      assert_equals(animation.playState, "finished",
        "animation playState update");
    }, "triggered animation with event listener can trigger after target is " +
       "removed.");
  </script>
  </body>

</html>
